requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.2.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pycparser==2.21
pytest==7.3.1
pytest-flask==1.2.0
pytest-mock==3.10.0
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==2.0.9
stripe==5.4.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
pytz~=2023.3
WTForms~=3.0.1
python-dateutil~=2.8.2
celery~=5.2.7
langchain~=0.0.147